completed questions

+++ shopping/order-service/src/main/java/com/patika/orderservice/model/Product.java
@@ -9,8 +9,8 @@ import lombok.*;
 @Builder
 public class Product {
 	
-	long id;
+	Long id;
 	String name;
-	double price;
+	Long price;
 }
